Move registration test to appropriate module
In order to have the least amount of work when removing automatic
registration, it's good to have its tests in one central location, so
the move is a no brainer.
Here's the explanation of what's happening. The `pandas` import and the
`timeindex` definition can be moved because the `timeindex` is only used
in the registration and `pandas` is only used in the timeindex
definition. The test was copied as is and the only thing that got added
to the test was the first line, where `Node.registry` is set to the
energy system created freshly for each test. This is important because
`Node.registry` is set to `None` for each test in order to not have
spurious registrations. Since the test uses some new things, these had
to be imported to, which explains the import changes. Last but not
least, since the moved test was the only test using `Transformer` this
import could be removed from "tests/basic_tests.py".
